chore: AttendMemberController 기능 분리
