+++ src/main/java/com/moa/controller/NoticeController.java
@@ -2,12 +2,9 @@ package com.moa.controller;
 
 import com.moa.dto.IdResponse;
 import com.moa.dto.notice.*;
-import com.moa.global.auth.model.JwtUser;
-import com.moa.service.AttendMemberService;
 import com.moa.service.NoticeService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import static org.springframework.http.HttpStatus.*;
@@ -18,7 +15,6 @@ import static org.springframework.http.HttpStatus.*;
 public class NoticeController {
 
     private final NoticeService noticeService;
-    private final AttendMemberService attendMemberService;
 
     @ResponseStatus(CREATED)
     @PostMapping
@@ -41,15 +37,4 @@ public class NoticeController {
         return noticeService.findAll(recruitmentId);
     }
 
-    @ResponseStatus(CREATED)
-    @PostMapping("/{noticeId}/vote/{attendance}")
-    public IdResponse voteAttendance(@PathVariable Long recruitmentId, @PathVariable Long noticeId, @PathVariable String attendance, @AuthenticationPrincipal JwtUser user){
-        VoteAttendanceRequest request = VoteAttendanceRequest.builder()
-                .recruitmentId(recruitmentId)
-                .noticeId(noticeId)
-                .attendance(attendance)
-                .userId(user.id())
-                .build();
-        return new IdResponse(attendMemberService.voteAttendance(request));
-    }
 }
